Remove redundant compiler and linker flags mistakenly added for HDF5
Strip out -I/usr/include, -L/usr/lib, and -L/usr/lib64
that wreak havoc when there are multiple versions of LALSuite
on the system. For example, this fixes an issue where LALSuite
libraries are linked against both /usr/lib64/liblal.so and
/path/to/local/prefix/lib/liblal.so.

FIXME: This is probably a bug in HDF5 itself or the HDF5
packaging. If HDF5 used pkg-config, then it would take care
of this automatically.